provide way to disable tmux config for ssh servers
The current tmux config isn't compatible with all versions of tmux.
When ssh'ing into servers, the bashrc will automatically launch tmux.
The version of tmux on the server might not be compatible with the tmux config.
This will cause the ssh session to start with tmux in an unpleasant state.
As a simple solution to the issue for now, bashrc will check for the existence
of a file `~/.tmux.conf-unsupported` to serve as a flag indicating that the host
does not have a suitable version of tmux for the tmux config.