add idempotent setup function for setting symlinks
The functions `ensure_inside` and `ensure_outside` also needed to be updated to
accomodate this function. During backups, if there was an existing backup at the
desired location that happened to be a symlink, it would get resolved to the
path to which it was pointing, which could be outside of the backup directory,
which would then lead to an error. This specific set of conditions would be a
huge problem since multiple similar invocations of `set_symlink` could reliably
reproduce it; you wouldn't be able to run the setup process twice! The solution
was to only resolve symlinks in the dirname portion of the path being checked
and not resolving symlinks in the tip of the path.