fix issue backing up to destinations with symlinks
Before performing an action that might overwrite a file, the setup functions
generally try to see if a file exists at the location in question so it can be
backed up. However, if the pre-existing file is a symlink pointing to a path
where nothing exists, then the test would fail to recognize that the symlink
exists because it would be resolving it to the target of the symlink which
itself does not exist. An additional '-L' test was added beside each relevant
'-e' test to solve the issue.