automatically install oh-my-posh in mysetup.sh
There was an issue that occurred when setting a symlink from the default
oh-my-posh theme's path to the `mo.omp.json` theme because after recursively
resolving the target of the symlink it would point inside the repo and trigger
an error. The 'only resolve the dirname' trick wasn't working because a symlink
along the path was pointing to a directory in the repo which contained the
ultimate target. Solving this seemed daunting and there didn't seem to be much
value in keeping the check so I just removed it from the `set_symlink` method to
keep things rolling.