deploy WT settings.json and BG image during setup
This required implementing a separate file deployment function for WSL->Windows
file system deployments. The main motivations for this were performance and
reliability. If Windows Terminal needs to look into the WSL file system every
time it launches or updates its settings, that may lead to bad user experience.
Furthermore, if there are file system incompatibility issues or if WSL is not
available for some reason, then this would cause issues in Windows Terminal.