Revert "remove personal configs"
This reverts commit ab35a7470cac82b3fe4178d0b4a14eb012bcdbd8.

The purpose of adding this commit to `main` and reverting it is to help with the
branching model moving forward. The plan is to commit the changes to my personal
configs to `main` and commit the changes to the generic toolkit to `blank` then
merge it into `main`. This workaround prevents that merge from removing my
personal configs in the `main` branch.