fix issues with GPG pinentry on Windows/WSL
It had been a while since I used the vscode GUI to create commits and
as I just tested it didn't work. I think it was caused by a change in
vscode that made the bashrc run underneath the hood for the GUI commit
operation and so the special pinentry variable was being set for
TERMINAL mode even though a terminal was not available. Checking that
bashrc is running inside a terminal before setting the special variable
seemed to fix the problem.